+++ src/Resources/contao/languages/de/tl_elolisten.php
@@ -25,5 +25,5 @@ $GLOBALS['TL_LANG']['tl_elolisten']['delete'] = array('Liste löschen', 'Liste %
 $GLOBALS['TL_LANG']['tl_elolisten']['toggle'] = array('Liste aktivieren/deaktivieren', 'Liste %s aktivieren/deaktivieren');
 $GLOBALS['TL_LANG']['tl_elolisten']['show'] = array('Listendetails anzeigen', 'Details des Listen %s anzeigen');
 // Zusätzliche Funktionen
-$GLOBALS['TL_LANG']['tl_elolisten']['import'] = array('FIDE-Daten in Liste %s importieren', 'Import von FIDE-Daten in die Liste %s');
-$GLOBALS['TL_LANG']['tl_elolisten']['importConfirm'] = 'Sollen die FIDE-Daten wirklich in diese Liste importiert werden? Vorhandene Spieler der Liste werden dabei überschrieben.';
+$GLOBALS['TL_LANG']['tl_elolisten']['import'] = array('FIDE-Daten importieren', 'FIDE-Daten in die Liste %s importieren');
+$GLOBALS['TL_LANG']['tl_elolisten']['importConfirm'] = 'Sollen die FIDE-Daten wirklich in diese Liste importiert werden? Vorhandene Spieler der Liste werden dabei überschrieben und Spieler, die in der Importdatei fehlen, gelöscht.';